+++ backend/src/main/java/com/best_umbrella/backend/repository/UtilizadorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UtilizadorRepository extends JpaRepository<Utilizador, Long> {
+public interface UtilizadorRepository extends JpaRepository<Utilizador, Integer> {
     Utilizador findByEmail(String email);
 }
