+++ backend/src/main/java/com/best_umbrella/backend/service/UtilizadorService.java
@@ -25,7 +25,7 @@ public class UtilizadorService {
         return utilizadorRepository.findAll();
     }
 
-    public Optional<Utilizador> findById(Long id) {
+    public Optional<Utilizador> findById(Integer id) {
         return utilizadorRepository.findById(id);
     }
